Refactor output to be more useful with large trees
If you have a large number of processes with long identical argvs,
the output tends to get crowded. This change will fold all the
processes with identical argvs into one row which is prefixed by
the list of PIDs. If there are more than five PIDs and verbose is not
set, the list is changed to "(<count> processes)" which is useful for
human consumption. For programmatic consumption, the list of PIDs can
be forced and will then ben "{PID1,PID2,PID3,...}". Note the difference
in parenthesis.