Simplify get_progargs()
The old scheme would delete whitespace at the edges of parameters. This
is a minor bug, but code also re-allocated argv several times and was a
lot more complex than needed.

The new version is not perfect, either, since it doesn't represent the
command line as used when starting the program --- there is a reason why
the proc file uses NULL bytes. Still, it's at least as accurate as the
old version and much more robust.