Fix systemd querying with newer versions
Starting sometime around v216, systemctl output changed from:

$ ./systemctl status 2846
sshd.service - OpenSSH Daemon
[...]

to:

$ ./systemctl status 2846
● sshd.service - OpenSSH Daemon
[...]

throwing off the parsing code. Ideally, we'd use a machine-readable
version of this information, but the status subcommand has no option for
that. The man page directs to the "show" subcommand, but that can not be
queried by PID.

newer systemd versions (well after 208) also have a "busctl" tool that
is closer to what we want --- but backward compatibility is just as
messy as this change.

Another possibility would be a raw DBUS query, but that is tricky to get
right, so we go with this patch and fix things for now.