Add fd_users tool
The purpose of this tool is similar to that of lib_users, but instead of
mapped files, it looks for open file descriptors that point to deleted
files. Since the logic of determining that is very different from the
mapped files case, it is a separate tool.

There still is some common code between the two tools, and that should
be factored out onto a separate library the tools use.